examples: silence narrow conversion warning/error
for some reason in Unix this only works if server is bound to <any>
(0.0.0.0 / [::])
but not if bound to specific interface
with IPv6 also the client must be bound to [::]

+++ examples/sockpuppet_udp_server.cpp
@@ -28,11 +28,10 @@ using namespace sockpuppet;
     if(bindAddress.IsV6()) {
       auto opt = ipv6_mreq{
         reinterpret_cast<sockaddr_in6 const *>(remoteAddress->impl->ForUdp().addr)->sin6_addr, // IP multicast address of group.
-        sock.LocalAddress().impl->LocalInterfaceIndex()
+        static_cast<unsigned int>(sock.LocalAddress().impl->LocalInterfaceIndex())
       };
       sock.impl->SetSockOpt(IPPROTO_IPV6, IPV6_ADD_MEMBERSHIP, reinterpret_cast<char const *>(&opt), sizeof(opt));
     } else {
-      // for IPv4 (IPv6 has its own ID and structure)
       auto opt = ip_mreq{
         reinterpret_cast<sockaddr_in const *>(remoteAddress->impl->ForUdp().addr)->sin_addr, // IP multicast address of group.
         reinterpret_cast<sockaddr_in const *>(sock.LocalAddress().impl->ForUdp().addr)->sin_addr // Local IP address of interface.
